requirements.txt
asn1crypto==1.3.0
azure-common==1.1.25
azure-core==1.6.0
azure-storage-blob==12.3.2
boto3==1.13.26
botocore==1.16.26
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
docutils==0.15.2
Flask==1.1.2
Flask-WTF==0.14.3
future==0.18.2
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
msrest==0.6.17
oauthlib==3.1.0
oscrypto==1.2.0
pycparser==2.20
pycryptodomex==3.9.8
PyJWT==1.7.1
pyngrok==4.1.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.2.8
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
